refactor: matcher モジュールに切り出す
