+++ src/components/SuggestionBox.tsx
@@ -1,7 +1,7 @@
-import Asearch from 'asearch';
 import { ComponentChild } from 'preact';
 import { useCallback, useEffect, useMemo, useState } from 'preact/hooks';
 import { JSXInternal } from 'preact/src/jsx';
+import { fuzzyMatcher } from '../lib/matcher';
 import { CursorPosition, Matcher } from '../types';
 import { PopupMenu } from './PopupMenu';
 import { QueryInput } from './SuggestionBox/QueryInput';
@@ -12,25 +12,6 @@ export type Item<T> = {
   searchableText: string;
   value: T;
 };
-
-export function matchItems<T>(query: string, items: Item<T>[]): Item<T>[] {
-  const maxAambig = Math.min(Math.floor(query.length / 4) + 1, 3);
-  const match = Asearch(` ${query} `); // 部分一致できるように、両端をスペースで囲む
-  // あいまい度の少ない項目から順に並べる
-  // 重複は除く
-  const pushedItems: Set<Item<T>> = new Set();
-  const result: Item<T>[] = [];
-  for (let ambig = 0; ambig <= maxAambig; ambig++) {
-    items.forEach((item) => {
-      if (!match(item.searchableText, ambig)) return;
-      if (pushedItems.has(item)) return;
-      pushedItems.add(item);
-      result.push(item);
-    });
-  }
-
-  return result;
-}
 
 export type SuggestionBoxProps<T> = {
   open: boolean;
@@ -49,7 +30,7 @@ export function SuggestionBox<T>({
   emptyMessage,
   items,
   cursorPosition,
-  matcher = matchItems,
+  matcher = fuzzyMatcher,
   onSelect,
   onSelectNonexistent,
   onClose,
